login validation & redirect to main if user is already logged
