+++ project/src/pages/login-page/login-page.tsx
@@ -1,30 +1,20 @@
-import { useRef, FormEvent } from 'react';
-import { useAppDispatch } from '../../hooks';
-import { loginAction } from '../../store/api-actions';
-import { AuthData } from '../../types/auth-data';
+import { Navigate } from 'react-router-dom';
+import { useAppSelector } from '../../hooks';
+import { getAuthorizationStatus } from '../../store/user-process/selectors';
+import { AuthorizationStatus, AppRoute } from '../../const';
 import Logo from '../../components/logo/logo';
+import LoginForm from '../../components/login-form/login-form';
 import Footer from '../../components/footer/footer';
 
 function LoginPage(): JSX.Element {
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
+  const authorizationStatus = useAppSelector(getAuthorizationStatus);
 
-  const dispatch = useAppDispatch();
+  if (authorizationStatus === AuthorizationStatus.Auth) {
+    return (
+      <Navigate to={AppRoute.Root} />
+    );
+  }
 
-  const onSubmit = (authData: AuthData) => {
-    dispatch(loginAction(authData));
-  };
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      onSubmit({
-        login: loginRef.current.value,
-        password: passwordRef.current.value,
-      });
-    }
-  };
   return (
     <div className="user-page">
       <header className="page-header user-page__head">
@@ -33,45 +23,10 @@ function LoginPage(): JSX.Element {
       </header>
 
       <div className="sign-in user-page__content">
-        <form
-          action="#"
-          className="sign-in__form"
-          onSubmit={handleSubmit}
-        >
-          <div className="sign-in__fields">
-            <div className="sign-in__field">
-              <input
-                ref={loginRef}
-                className="sign-in__input"
-                type="email" placeholder="Email address"
-                name="user-email"
-                id="user-email"
-              />
-              <label className="sign-in__label visually-hidden" htmlFor="user-email">Email address</label>
-            </div>
-            <div className="sign-in__field">
-              <input
-                ref={passwordRef}
-                className="sign-in__input" type="password"
-                placeholder="Password" name="user-password"
-                id="user-password"
-              />
-              <label className="sign-in__label visually-hidden" htmlFor="user-password">Password</label>
-            </div>
-          </div>
-          <div className="sign-in__submit">
-            <button
-              className="sign-in__btn"
-              type="submit"
-            >
-              Sign in
-            </button>
-          </div>
-        </form>
+        <LoginForm />
       </div>
 
       < Footer />
-
     </div>
   );
 }
